refactor(api): move create function from tag controller
in src/organizational-entities
