+++ api/lib/application/tags/tag-controller.js
@@ -1,12 +1,6 @@
 import * as tagSerializer from '../../../src/organizational-entities/infrastructure/serializers/jsonapi/tag-serializer.js';
 import { usecases } from '../../domain/usecases/index.js';
 
-const create = async function (request, h, dependencies = { tagSerializer }) {
-  const tagName = request.payload.data.attributes['name'].toUpperCase();
-  const createdTag = await usecases.createTag({ tagName });
-  return h.response(dependencies.tagSerializer.serialize(createdTag)).created();
-};
-
 const findAllTags = async function (request, h, dependencies = { tagSerializer }) {
   const organizationsTags = await usecases.findAllTags();
   return dependencies.tagSerializer.serialize(organizationsTags);
@@ -18,6 +12,6 @@ const getRecentlyUsedTags = async function (request, h, dependencies = { tagSeri
   return dependencies.tagSerializer.serialize(recentlyUsedTags);
 };
 
-const tagController = { create, findAllTags, getRecentlyUsedTags };
+const tagController = { findAllTags, getRecentlyUsedTags };
 
 export { tagController };
