+++ api/tests/unit/application/tags/tag-controller_test.js
@@ -3,29 +3,6 @@ import { usecases } from '../../../../lib/domain/usecases/index.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../test-helper.js';
 
 describe('Unit | Application | Tags | tag-controller', function () {
-  describe('#create', function () {
-    it('should return the created tag', async function () {
-      // given
-      const createdTag = domainBuilder.buildTag({ id: 1, name: 'TAG1' });
-      const serializedTag = Symbol('a serialized tag');
-
-      sinon.stub(usecases, 'createTag').resolves(createdTag);
-      const tagSerializer = {
-        serialize: sinon.stub(),
-      };
-      tagSerializer.serialize.withArgs(createdTag).returns(serializedTag);
-
-      const request = { payload: { data: { attributes: { name: 'tag1' } } } };
-
-      // when
-      const result = await tagController.create(request, hFake, { tagSerializer });
-
-      // then
-      expect(result.statusCode).to.be.equal(201);
-      expect(result.source).to.be.equal(serializedTag);
-    });
-  });
-
   describe('#findAllTags', function () {
     it('should call findAllTags usecase and tag serializer', async function () {
       // given
